Integrated templates, automated adc_floats with a scan field in read_voltage record and found out that reading even at .1 second is no problem for multiple channels. So the reason why the boxes sometimes crash seems to be the caput command respectively the fact that it is called to often and the amount of connections is to much for the box - at least I guess so from what I know now. Thanks to the scan field, gathering the voltage values now happens automatically when you start epics and it is only necessary to use the caget command to get the value - the caput command is not needed. The Command currently looks like that: caget CB:HV:BOX:<boxnumber>:<levelnumber>:<channelnumber>:read_volt
Moreover just some additional notes for myself about EPICS so I don't forget them (whoever reads this has to judge on their own if these informations are valuable or not ;-)): EPICS stores values in its own 'memory'. This is why you get only the number of the level and channel back instead of the desired voltage when you 'manually' use the caput and caget command to fast after each other. The caput command orders the box to give back the voltage value to EPICS, the caget command 'forwards' the value to eg the terminal. However, until the voltage value from the box comes in, the command given to box is in the same memory. When the caget command is processed before the voltage value from the box reaches EPICS, the currently stored value (= the command) is forwarded. So when you want to know eg the voltage of box 19, level 1, channel 1 and caget is sent to fast after the caput command, the displayed value is '1 1' instead of something like '1234.567'.