After several attempts - finally - this version is the first one to work properly. Still I need to find out if all the changes I made are really necessary, but at least these files now do the trick. Most import - by using wireshark I found out, that after every command the box also returns (meant is a full command like the entire response to the 'eemem print' command, not after every line) the hv box adds a '>' sign. I now use this as terminator when looking for the end of the input after longer commands and together with an extended wait for replytimeouts the record now reads the entire response and it no loner mixes up with other packages sent from the box. Another important fact is, that the maximum package size the box can send is 582 bytes. This was a problem as epics thought after receiving the first string with 582 bytes, this is all the input from the box. Indeed there was at least one or even more packages also belonging to the response (e.g. the the input from the box after sending the 'eemem print' command consists of three packages). Now epics 'knows' it has to wait for more to come.
I also added the functions to read out and set values within the eemem, though this is still a little experimental at the moment. The read function works as follows: The values in the eemem are stored in order of their last change - the value that was changed last is at the bottom. As there is no fixed line, a specific eemem value can be found in, I used a trick mentioned in the streamDevice docu. For every eemem value (currently as separate records, but I'm thinking of creating a template for that) there is a record in 'I/O Intr' scan mode. By using this scan mode these records get a copy of the entire input from the box. They don't ever cause a mismatch error handler, but 'wait' until they receive a string that matches their 'in' protocol. Then the record is processed and starts again looking for the next matching input. After talking to Mr. Corell I know, that though it is possible to delete a variable from the eemem without adding it afterwards with another value this is not intended. So the number of lines the box sends when prompted using 'eemem print' is fixed. As the passivly scanning records using 'I/O Intr' get copies of everything anyway, the 'challange' is to get rid of the original input. Knowing the number of lines this is fairly easy - all I had to do was either wait for the '>' symbol or tell the system to wait for a specific amount of lines before continuing - I have to figure out, whether one way is better than the other - still both seem to work for the moment.

In a nutshell - currently - the inputs from the box are no longer mixed up and though the Replytimeout setting is slowing the actively scanning records a bit it least for now seems to work.